Abstract overlay event component
