+++ apps/website/src/pages/stream/overlay.tsx
@@ -1,18 +1,9 @@
 import { type NextPage } from "next";
-import Image from "next/image";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Transition } from "@headlessui/react";
-import { keepPreviousData } from "@tanstack/react-query";
+import { useEffect, useRef, useState } from "react";
 
-import { trpc } from "@/utils/trpc";
-import {
-  DATETIME_ALVEUS_ZONE,
-  formatDateTimeParts,
-  formatDateTimeRelative,
-} from "@/utils/datetime";
-import { getFormattedTitle, twitchChannels } from "@/data/calendar-events";
+import { DATETIME_ALVEUS_ZONE, formatDateTimeParts } from "@/utils/datetime";
 
-import logoImage from "@/assets/logo.png";
+import Event from "@/components/overlay/Event";
 
 import { type WeatherResponse } from "../api/stream/weather";
 
@@ -106,56 +97,6 @@ const OverlayPage: NextPage = () => {
     return () => clearTimeout(weatherStaleTimer.current ?? undefined);
   }, [weather]);
 
-  // Set the range for upcoming events to the next 3 days
-  // Refresh every 60s
-  const [upcomingRange, setUpcomingRange] = useState<[Date, Date]>();
-  const upcomingRangeInterval = useRef<NodeJS.Timeout>(null);
-  useEffect(() => {
-    const updateRange = () => {
-      const now = new Date();
-      const next = new Date(now);
-      next.setDate(next.getDate() + 3);
-      setUpcomingRange([now, next]);
-    };
-
-    updateRange();
-    upcomingRangeInterval.current = setInterval(updateRange, 60 * 1000);
-    return () => clearInterval(upcomingRangeInterval.current ?? undefined);
-  }, []);
-
-  // Get the upcoming events and the first event ID
-  // Refresh when the range changes
-  const { data: events } = trpc.calendarEvents.getCalendarEvents.useQuery(
-    { start: upcomingRange?.[0], end: upcomingRange?.[1] },
-    { enabled: upcomingRange !== undefined, placeholderData: keepPreviousData },
-  );
-  const firstEventId = useMemo(
-    () => events?.find(twitchChannels.alveus.filter)?.id,
-    [events],
-  );
-
-  // If we have an upcoming event, swap socials with it
-  // Swap every 60s
-  const [visibleEventId, setVisibleEventId] = useState<string>();
-  const eventInterval = useRef<NodeJS.Timeout>(null);
-  useEffect(() => {
-    if (!firstEventId) {
-      setVisibleEventId(undefined);
-      return;
-    }
-
-    const swapEvent = () =>
-      setVisibleEventId((prev) => (prev ? undefined : firstEventId));
-    swapEvent();
-    eventInterval.current = setInterval(swapEvent, 60 * 1000);
-    return () => clearInterval(eventInterval.current ?? undefined);
-  }, [firstEventId]);
-  const event = useMemo(
-    () =>
-      visibleEventId && events?.find((event) => event.id === visibleEventId),
-    [visibleEventId, events],
-  );
-
   // This can be a client-side only page
   if (!time) return null;
 
@@ -172,46 +113,7 @@ const OverlayPage: NextPage = () => {
         )}
       </div>
 
-      <Transition show={event !== undefined}>
-        <div className="text-stroke absolute bottom-2 left-2 font-bold text-white transition-opacity data-[closed]:opacity-0 data-[enter]:duration-700 data-[leave]:duration-300">
-          <p>Upcoming:</p>
-
-          {event && (
-            <>
-              <p className="text-xl">
-                {getFormattedTitle(event, twitchChannels.alveus.username, 30)}
-              </p>
-              <p className="text-xl">
-                {formatDateTimeRelative(
-                  event.startAt,
-                  {
-                    style: "long",
-                    time: event.hasTime ? "minutes" : undefined,
-                    timezone: event.hasTime,
-                  },
-                  { zone: DATETIME_ALVEUS_ZONE },
-                )}
-              </p>
-            </>
-          )}
-        </div>
-      </Transition>
-
-      <Transition show={event === undefined}>
-        <div className="absolute bottom-2 left-2 flex items-center gap-2 transition-opacity data-[closed]:opacity-0 data-[enter]:duration-700 data-[leave]:duration-300">
-          <Image
-            src={logoImage}
-            alt=""
-            height={64}
-            className="h-16 w-auto opacity-75 brightness-150 contrast-125 drop-shadow grayscale"
-          />
-
-          <div className="text-stroke text-xl font-bold text-white">
-            <p>alveussanctuary.org</p>
-            <p>@alveussanctuary</p>
-          </div>
-        </div>
-      </Transition>
+      <Event className="absolute bottom-2 left-2" />
 
       <div className="absolute bottom-0 right-0 grid grid-cols-12">
         {time.code.map((bit, idx) => (
